Add smallrye implementation of MetricRegistry
Since the MicroProfile library is just an API, a concrete instance of
MetricRegistry can't be instantiated inside MicroProfileMetricsFactory.
To allow a Factory class to be configured, it will have to have a
constructor that only tkaes an instance of DriverContext, so an instance
of MetricRegistry will have to be created within the Factory.